refactor(folder structure): create windows and animations folders
