+++ src/Components/Animations/Bomb/Bomb.scss
