+++ src/Components/Windows/CharacterWindow/CharacterWindow.scss
