+++ src/Components/Windows/CharacterWindow/HealthBar/HealthBar.scss
