+++ src/Components/Windows/LoginWindow/logic.js
@@ -1,6 +1,6 @@
 import firebase from 'firebase/app';
 import 'firebase/auth';
-import firebaseConfig from '../../../firebaseConfig';
+import firebaseConfig from '../../../../firebaseConfig';
 
 firebase.initializeApp(firebaseConfig);
