+++ src/Components/Windows/MonsterCounterWindow/MonsterCounterWindow.scss
