+++ src/Components/Windows/MonsterCounterWindow/images/monster-dead-head.png
