+++ src/Components/Windows/ScoreWindow/ScoreItem/ScoreItem.scss
