+++ src/Components/Windows/ScoreWindow/ScoreWindow.scss
