+++ src/Components/Windows/SpellWindow/SpellWindow.scss
